Fix 1 Rubocop offense by refactoring the whole test suite.
