--- spec/extensions/jekyll/site_spec.rb
@@ -1,5 +1,0 @@
-require './spec/helper'
-
-describe Jekyll::Site do
-  specify { Jekyll::Site.include?(JekyllSiteExtensions).must_equal(true) }
-end
