+++ spec/helper.rb
@@ -8,10 +8,6 @@ require 'minitest/pride'
 require './spec/helpers/extensions/ruby/module'
 require 'jekyll_asset_pipeline'
 
-# rubocop:disable Style/MixinUsage
-include JekyllAssetPipeline
-# rubocop:enable Style/MixinUsage
-
 module MiniTest
   class Spec
     def source_path
