+++ spec/jekyll_asset_pipeline/extensions/jekyll/site_spec.rb
@@ -1,0 +1,7 @@
+require './spec/helper'
+
+module JekyllAssetPipeline
+  describe Jekyll::Site do
+    specify { Jekyll::Site.include?(JekyllSiteExtensions).must_equal(true) }
+  end
+end
