+++ spec/jekyll_asset_pipeline/extensions/liquid/asset_tag_spec.rb
@@ -1,0 +1,10 @@
+require './spec/helper'
+
+module JekyllAssetPipeline
+  describe AssetTag do
+    specify do
+      AssetTag.extend?(LiquidBlockExtensions::ClassMethods).must_equal(true)
+      AssetTag.include?(LiquidBlockExtensions).must_equal(true)
+    end
+  end
+end
