+++ spec/jekyll_asset_pipeline/extensions/liquid/asset_tags/java_script_asset_tag_spec.rb
@@ -1,0 +1,17 @@
+require './spec/helper'
+
+module JekyllAssetPipeline
+  describe JavaScriptAssetTag do
+    specify do
+      JavaScriptAssetTag.tag_name.must_equal('javascript_asset_tag')
+      JavaScriptAssetTag.output_type.must_equal('.js')
+      (JavaScriptAssetTag.superclass == JekyllAssetPipeline::AssetTag)
+        .must_equal(true)
+    end
+
+    it 'registers tag with Liquid' do
+      ::Liquid::Template.tags[JekyllAssetPipeline::JavaScriptAssetTag.tag_name]
+                        .must_equal(JekyllAssetPipeline::JavaScriptAssetTag)
+    end
+  end
+end
