+++ spec/jekyll_asset_pipeline/version_spec.rb
@@ -1,0 +1,11 @@
+require './spec/helper'
+
+module JekyllAssetPipeline
+  describe VERSION do
+    subject { JekyllAssetPipeline::VERSION }
+
+    it 'returns a string' do
+      subject.must_be_instance_of(String)
+    end
+  end
+end
